move first lunch text code to new file
